packages/python/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "quanttide-devops"
version = "0.1.0-beta.1"
description = "Python SDK for Specification of QuantTide DevOps"
authors = ["QuantTide Inc. <opensource@quanttide.com>"]
readme = "README.md"
packages = [{include = "quanttide_devops"}]

[tool.poetry.dependencies]
python = "^3.9"
gitpython = "^3.1.40"
semantic-version = "^2.10.0"
pygithub = "^2.1.1"
coding-devops-sdk-py = "^0.1.2"
pydantic = "^2.5.2"
typer = "^0.9.0"

[tool.poetry.scripts]
qtdevops = "quanttide_devops.cli:app"
